+++ packages/doggy-nav-admin/src/constants/index.ts
@@ -3,3 +3,16 @@ export const TOKEN = 'token';
 export const CURRENT_USER = 'currentUser';
 export const GLOBAL_CATEGORY_ID = '4bvirtualcb9ff050738cc16';
 export const GLOBAL_CATEGORY_NAME = '全局根分类';
+export const CATEGORY_NAME_ZH_MAP: Record<string, string> = {
+  Overview: '概览',
+  'Tech & Dev': '技术与开发',
+  Design: '设计',
+  Business: '商业',
+  Learn: '学习',
+  Social: '社交',
+  Media: '媒体',
+  Tools: '工具',
+  Shop: '购物',
+  Life: '生活',
+  Science: '科学',
+};